Fixed timer overflow
- Changed timer from unsigned int to unsigned long to resolve overflow issue resulting in games being lost as soon as they start

+++ SimonStickTest/SimonStickTest.ino
@@ -57,7 +57,7 @@ bool getPlayerIn()
   for(byte count{0};count < 4;++count)
   {
     byte tilt{};
-    unsigned int timer{millis() + timeLimit};
+    unsigned long timer{millis() + timeLimit};
     while(true)
     {
       if(millis() > timer)
